Add buddy allocator for physical page management
Introduce a configurable buddy allocator that manages physical pages with
page descriptors, split/coalesce support, aligned allocations, and memory
accounting.

The allocator reuses the existing blueos_infra intrusive list
implementation for free lists instead of adding a separate linked-list
implementation.

Wire the buddy allocator into the kernel allocation path so slab can grow
from buddy-managed page pools and page-sized-or-larger allocations can be
served directly by buddy.

Add physical DRAM range metadata for qemu_riscv32, qemu_riscv64,
qemu_virt64_aarch64, and rk3568, and enable the allocator in
their defconfigs.

Move kernel physical/virtual address translation to shared mm helpers,
update AArch64 linear mapping setup to map the configured DRAM range, and
use the shared translation helpers in virtio.

Add buddy unit tests covering initialization, allocation/free, exhaustion,
split/coalesce, alignment, MAX_ORDER boundaries, stress sequences,
and order calculation.

+++ kernel/src/boards/qemu_virt64_aarch64/init.rs
@@ -1,4 +1,4 @@
-// Copyright (c) 2025 vivo Mobile Communication Co., Ltd.
+// Copyright (c) 2026 vivo Mobile Communication Co., Ltd.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -90,7 +90,7 @@ crate::define_peripheral! {
      )),
 }
 
-pub const DRAM_BASE: u64 = mmu::kernel_phys_to_virt(0x4000_0000);
+pub const DRAM_BASE: usize = mmu::kernel_phys_to_virt(0x4000_0000);
 
 crate::define_pin_states!(None);
